+++ src/test/java/flab/nutridiary/SpringBootTestSupport.java
@@ -1,7 +1,5 @@
 package flab.nutridiary;
 
-import com.amazonaws.services.s3.AmazonS3;
-import flab.nutridiary.commom.config.S3Config;
 import flab.nutridiary.commom.file.FileStoreService;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -12,12 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 @SpringBootTest
 public abstract class SpringBootTestSupport {
-
-    @MockBean
-    protected S3Config s3Config;
-
-    @MockBean
-    protected AmazonS3 amazonS3;
 
     @MockBean
     protected FileStoreService fileStoreService;
