+++ src/test/java/flab/nutridiary/TestContainerSupport.java
@@ -1,7 +1,5 @@
 package flab.nutridiary;
 
-import com.amazonaws.services.s3.AmazonS3;
-import flab.nutridiary.commom.config.S3Config;
 import flab.nutridiary.commom.file.FileStoreService;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -18,12 +16,6 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @SpringBootTest
 public abstract class TestContainerSupport {
     static final MySQLContainer<?> mysqlContainer;
-
-    @MockBean
-    protected S3Config s3Config;
-
-    @MockBean
-    protected AmazonS3 amazonS3;
 
     @MockBean
     protected FileStoreService fileStoreService;
